Skips apps that are already running/closed for opening/closing
* Uses the unix process status command to check if each app is running
or not. If it is, and the user is asking to open it, it will be skipped.
Likewise, if the app is already not running, and the program wants to
close it, that command will be skipped.

* It's imperfect, because `grep`ping for an app by name returns
background processes it might keep running (looking at you Google
Chrome).

* I don't think this is actually faster, but I like the idea of avoiding
redundancy.

* Also refactors the open/close actions into functions, and uses the
`-SIGINT` flag when closing because it is more graceful, I think.